ocaml/ArrayCreationTest.testBooleanArrayCreationNoInitializer passes
Went through the codebase to get this single test to pass. It
translates

boolean[] foo = new boolean[3];
to
let foo : bool array = Array.make 3 true

Still much of the codebase I do not understand. Commits will continue
to be very sloppy for a while as I get an understanding of the codebase.