All primitive array tests pass
Two tests are marked 'FIXTEST' for now: the object array, and
the multidimensional array. The multidimensional array will work
correctly right now, eg 'let foo = [| [| 4 |] |]' but the spirit
of their test seems to be about multidimensional numerical arrays
which we should map to BigArray in OCaml. I will return to these
later.

Decided to change byte to int instead of char. This was because
byte literals in java are ints. I did not want to deal with, at this
stage, converting.

Also, there are a couple of tests here that use java "1.3f" float
notation. I skipped removing the 'f' for now.